+++ service/policy/db/key_access_server_registry.go
@@ -601,14 +601,14 @@ func validateUnsafeUpdateKey(existing *policy.KasKey, r *unsafe.UnsafeUpdateKeyR
 	params := unsafeUpdateKeyParams{
 		ID: r.GetId(),
 	}
-	newKeyMode := pgtypeInt4(int32(r.GetKeyMode()), true)
+	newKeyMode := pgtypeInt4(int32(r.GetTargetKeyMode()), true)
 	newProviderConfiguration := pgtypeUUID(r.GetProviderConfigId())
 
 	if existingMode != policy.KeyMode_KEY_MODE_PUBLIC_KEY_ONLY && existingMode != policy.KeyMode_KEY_MODE_REMOTE {
 		return params, db.ErrUnsafeUpdateKeyExistingModeUnsupported
 	}
 
-	switch r.GetKeyMode() {
+	switch r.GetTargetKeyMode() {
 	case policy.KeyMode_KEY_MODE_REMOTE:
 		if !newProviderConfiguration.Valid {
 			return params, db.ErrUnsafeUpdateKeyProviderConfigRequired
@@ -621,13 +621,17 @@ func validateUnsafeUpdateKey(existing *policy.KasKey, r *unsafe.UnsafeUpdateKeyR
 		}
 		params.KeyMode = newKeyMode
 	case policy.KeyMode_KEY_MODE_UNSPECIFIED:
-		if existingMode != policy.KeyMode_KEY_MODE_REMOTE {
-			return params, db.ErrUnsafeUpdateKeyProviderConfigExistingMode
-		}
 		if !newProviderConfiguration.Valid {
 			return params, db.ErrUnsafeUpdateKeyProviderConfigRequired
 		}
+		if existingMode != policy.KeyMode_KEY_MODE_REMOTE {
+			return params, db.ErrUnsafeUpdateKeyProviderConfigExistingMode
+		}
 		params.ProviderConfigID = newProviderConfiguration
+	case policy.KeyMode_KEY_MODE_CONFIG_ROOT_KEY, policy.KeyMode_KEY_MODE_PROVIDER_ROOT_KEY:
+		fallthrough
+	default:
+		return params, fmt.Errorf("%w: %s", db.ErrUnsafeUpdateKeyTargetModeUnsupported, r.GetTargetKeyMode())
 	}
 
 	return params, nil